// js/src/types/RecordMapping.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AmountField = { "FromColumn": { col: number, invert: boolean, } } | { "FromCreditDebitColumns": { first: number, invert_first: boolean, second: number, invert_second: boolean, } };

export type DateField = { "FromColumn": { col: number, tz: TZ, } };

export type RecordMapping = { transaction_date: DateField, transaction_time: TimeField, description: TextField, amount: AmountField, };

export type TZ = "Local" | "UTC";

export type TextField = { "FromColumn": { col: number, } };

export type TimeField = { "FromColumn": { col: number, tz: TZ, } } | "Empty";